willans formula in rust, float precision problem
Float precision problem limits the usage in rust, maybe there is a way to work around it, i dunno. The problem is in the prime_checker function,
the product should become 1.0 if the input is a prime and it doesn't after the prime 17, it becomes 0.99999 and floor() turns it into 0 -> not prime even though it is prime.